+++ services/team-service/src/main/java/com/collab/teamservice/api/dto/TeamMemberView.java
@@ -4,5 +4,5 @@ public record TeamMemberView(
     String userId,
     String fullName,
     String memberRole,
-    Integer contributionPercent
+    Double contributionPercent
 ) {}
